backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
Flask-Cors = "^3.0.10"
flask-mysqldb = "^1.0.1"
python-dotenv = "^0.21.0"
pytest = "^7.2.1"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
